Make events async
Kind of an experiment--i'm using the dispatch/cofx/fx parts of re-frame
on the server side. So a call to the API will dispatch (asynchronously)
a (series of) events. Currently i just have two effects handlers, one
of which does almost ALL the effects--update db, update google, and
sync to beeminder. The other effect handler is just to send slack
messages. Ideally I should split my effects handlers so they are as
stupid-simple as possible, but this works for now and is a more minimal
change.