vague steps in the direction of graph improvements
C2 is deprecated but has a lot of inspiration. So I'm basically pulling
chunks of it into this repo, for axes, tick selection (in the future),
math, etc.

For now I added axes to the main dashboard graph and added a
(day-binned) histogram when you've entered a query.

It's definitely hacky and not performant, but it works!