Use previous local-time when given not parse-able
Someday I should add some logic to try to guess the date better. But
this works okay--just use the last local offset we have. Only time it
will fail *really* badly is if you do something like travel across 12
time zones, and you always respond with tons of tags to tagtime (so it
uses the shorter date format). And it should at least be an improvement
over the previous behavior (crash catastrophically when we hit an
unparseable date).