don't use session auth on sente /chsk handler
Sente really wants us to use sessions + CSRF tokens. But I'd like to
avoid that whole issue by just requiring authentication to be explicitly
sent with every request the user makes (instead of automatically being
sent by the browser). So basically allow the API token to be passed in
one of two ways: as an Authentication header, or by an "authentication"
param. Either way, no CSRF token is necessary.